% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmSurv.R
\name{ImmSurv}
\alias{ImmSurv}
\title{ImmSurv}
\usage{
ImmSurv(
  Data,
  ID = "ID",
  Status = "Status",
  Survival = "Survival",
  Stage = "Stage",
  Variable1 = "Variable1",
  Variable1Unit = "Variable1Unit",
  Variable2 = "Variable2",
  Variable2Unit = "Variable2Unit"
)
}
\description{
ImmSurv
}
