% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCut.R
\name{rCut}
\alias{rCut}
\title{rCut
This produces cut offs for each variable using the maximal rank statistics method.}
\usage{
rCut(
  Data,
  CSS,
  OS,
  PFS,
  ID,
  Survival,
  SurvivalStatus,
  Progression,
  ProgressionStatus,
  Variables
)
}
\description{
rCut
This produces cut offs for each variable using the maximal rank statistics method.
}
