% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCut.R
\name{rCut}
\alias{rCut}
\title{rCut
This produces cut offs for each variable using the maximal rank statistics method.}
\usage{
rCut(
  Data,
  CSS,
  OS,
  DFS,
  RFS,
  MISC,
  minprop = "0.1",
  PlotPalette = "SPSS",
  ID,
  Survival = "Survival",
  SurvivalStatus = "SurvivalStatus",
  DiseaseFree = "DiseaseFree",
  DiseaseFreeStatus = "DiseaseFreeStatus",
  Recurrence = "Recurrence",
  RecurrenceStatus = "RecurrenceStatus",
  Miscellaneous = "Miscellaneous",
  MiscellaneousStatus = "MiscellaneousStatus",
  Variables
)
}
\description{
rCut
This produces cut offs for each variable using the maximal rank statistics method.
}
